Lift up feature flags to libs/common
