+++ apps/web/src/utils/flags.ts
@@ -1,3 +1,11 @@
+// required to avoid linting errors when there are no feature flags
+/* eslint-disable @typescript-eslint/ban-types */
+
+import {
+  flagEnabled as baseFlagEnabled,
+  devFlagEnabled as baseDevFlagEnabled,
+} from "@bitwarden/common/misc/flags";
+
 export type Flags = {
   showTrial?: boolean;
 };
@@ -5,15 +13,13 @@ export type Flags = {
 export type FlagName = keyof Flags;
 
 export function flagEnabled(flag: FlagName): boolean {
-  return flags()[flag] == null || flags()[flag];
+  return baseFlagEnabled<Flags>(flag);
 }
 
-function flags(): Flags {
-  const envFlags = process.env.FLAGS as string | Flags;
+export type DevFlags = {};
 
-  if (typeof envFlags === "string") {
-    return JSON.parse(envFlags) as Flags;
-  } else {
-    return envFlags as Flags;
-  }
+export type DevFlagName = keyof DevFlags;
+
+export function devFlagEnabled(flag: DevFlagName) {
+  return baseDevFlagEnabled<DevFlags>(flag);
 }
