+++ libs/common/src/misc/flags.ts
@@ -1,0 +1,35 @@
+function getFlags<T>(envFlags: string | T): T {
+  if (typeof envFlags === "string") {
+    return JSON.parse(envFlags) as T;
+  } else {
+    return envFlags as T;
+  }
+}
+
+/**
+ * Gets the value of a feature flag from environment.
+ * All flags default to "on" (true)
+ * @param flag The name of the feature flag to check
+ * @returns The value of the flag
+ */
+export function flagEnabled<Flags extends Record<string, boolean>>(flag: keyof Flags): boolean {
+  const flags = getFlags<Flags>(process.env.FLAGS);
+  return flags[flag] == null || flags[flag];
+}
+
+/**
+ * Gets the value of a dev flag from environment.
+ * Will always return false unless in development.
+ * @param flag The name of the dev flag to check
+ * @returns The value of the flag
+ */
+export function devFlagEnabled<DevFlags extends Record<string, boolean>>(
+  flag: keyof DevFlags
+): boolean {
+  if (process.env.ENV !== "development") {
+    return false;
+  }
+
+  const devFlags = getFlags<DevFlags>(process.env.DEV_FLAGS);
+  return devFlags[flag] == null || devFlags[flag];
+}
